Split next-build from next-dev
