+++ Cargo.toml
@@ -4,6 +4,7 @@ resolver = "2"
 members = [
   "crates/auto-hash-map",
   "crates/next-binding",
+  "crates/next-build",
   "crates/next-core",
   "crates/next-dev",
   "crates/next-dev-tests",
