+++ crates/next-build/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "next-build"
+version = "0.1.0"
+description = "TBD"
+license = "MPL-2.0"
+edition = "2021"
+autobenches = false
+
+[features]
+next-font-local = ["next-core/next-font-local"]
+native-tls = ["next-core/native-tls"]
+rustls-tls = ["next-core/rustls-tls"]
+
+[dependencies]
+anyhow = "1.0.47"
+clap = { version = "4.0.18", features = ["derive", "env"], optional = true }
+mime = "0.3.16"
+next-core = { path = "../next-core" }
+owo-colors = "3"
+serde = "1.0.136"
+tokio = { version = "1.21.2", features = ["full"] }
+turbo-malloc = { path = "../turbo-malloc", default-features = false }
+turbo-tasks = { path = "../turbo-tasks" }
+turbo-tasks-fs = { path = "../turbo-tasks-fs" }
+turbo-tasks-memory = { path = "../turbo-tasks-memory" }
+turbopack-cli-utils = { path = "../turbopack-cli-utils" }
+turbopack-core = { path = "../turbopack-core" }
+turbopack-dev-server = { path = "../turbopack-dev-server" }
+turbopack-node = { path = "../turbopack-node" }
+
+[build-dependencies]
+turbo-tasks-build = { path = "../turbo-tasks-build" }
+vergen = { version = "7.3.2", default-features = false, features = [
+  "cargo",
+  "build",
+] }
