+++ crates/next-dev/Cargo.toml
@@ -34,9 +34,9 @@ tokio_console = [
 ]
 profile = []
 custom_allocator = ["turbo-malloc/custom_allocator"]
-next-font-local = ["next-core/next-font-local"]
-native-tls = ["next-core/native-tls"]
-rustls-tls = ["next-core/rustls-tls"]
+next-font-local = ["next-build/next-font-local"]
+native-tls = ["next-build/native-tls"]
+rustls-tls = ["next-build/rustls-tls"]
 
 [dependencies]
 anyhow = { version = "1.0.47", features = ["backtrace"] }
@@ -45,18 +45,16 @@ console-subscriber = { version = "0.1.8", optional = true }
 dunce = "1.0.3"
 futures = "0.3.25"
 mime = "0.3.16"
-next-core = { path = "../next-core" }
+next-build = { path = "../next-build" }
 owo-colors = "3"
 serde = "1.0.136"
 tokio = { version = "1.21.2", features = ["full"] }
 turbo-malloc = { path = "../turbo-malloc", default-features = false }
 turbo-tasks = { path = "../turbo-tasks" }
-turbo-tasks-fs = { path = "../turbo-tasks-fs" }
 turbo-tasks-memory = { path = "../turbo-tasks-memory" }
 turbopack-cli-utils = { path = "../turbopack-cli-utils" }
 turbopack-core = { path = "../turbopack-core" }
 turbopack-dev-server = { path = "../turbopack-dev-server" }
-turbopack-node = { path = "../turbopack-node" }
 webbrowser = "0.7.1"
 
 [dev-dependencies]
